Centralizando os testes e unificando os helpers
